Changed permissions collection to users
This represents a mindshift in the use of the structure.
The structure will now also contain a list of registered inputs to the database - making it fundamentally intertwined with the other parts, not only as a permissions keeper.
This change simplifies life, as the user can now have multiple meta-values all in one document - I don't have to add complexity for "input registry" and such