Modifying structure of user in progress
Originally I had planned outputs to be separate collection, but I already have the input register in the user collection. It only makes sense that the output register is also in the user object.
There will no longer be any real separation between inputs and outputs - the data associated with them will be all kept in the same place - in the dataset collection.
The register structure is also getting a big overhaul (once again).