Started migration to cryptomongo
The old method was limited to one computer, and was very much sequential.
The new method replaces cryptoServer with a "cryptomongo" server which manages the entire database.
The advantage of this method is that it allows future expansion into replica sets, and it further allows different parts of the database to be distributed to many computers (once auth is implemented, at least)