Splitting crypto to new module
The crypto code is useful for several databases - not only mongo. In the interest of not repeating code, I am splitting everything related to cryptography and data-safety into a new module called "cryptobase".

Cryptobase will make it really easy to create applications which use LUKS containers for data at rest encryption, and which use stunnel for encrypting communication between processes.