Connector client working
Still need to check if it works for kafka, but no reason not to (need internet to figure out the API).
Once I make sure it works for kafka, I think I am at v0.2. I just need internet to figure out if kafka is working as wanted. ANd I need internet to remind me how to tag a commit. So I will move forward with what I have now, and tag the first branch with internet as 0.2