Moved more stuff around
I am getting familiar with my old code again - and it is increasingly clear that I can do it again in a better way now. The entire mongoObject should be an automatic general method for accessing the database. I need to wrap dict and array to work correctly with mongoDB.

It sounds like a lot of work, but it has to be done. It also looks like I will end up entirely rewriting my old code.