Changed default partitions to 1
I have yet to figure out how to create partitions during runtime, and forcing messages to go by partition (multiple partitions give out of order messages. Partitions would be per-user or something.
This is also the v0.2 commit. The server is basically functional, so a version increment is in order.