kafka partition limits
Added partiiton limit bytes to kafka. This property will need to be further modified to take into account database file size as well as partition number, and the different topics.

Lastly, the configs for databases has numbers as strings. The config file writer does conversion already, so it is simpler to keep as ints.